voyagarの設定

+++ database/migrations/2021_07_21_001351_create_likes_table.php
@@ -15,9 +15,9 @@ class CreateLikesTable extends Migration
     {
         Schema::create('likes', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id')->unsigned()->nullable();
+            $table->bigInteger('user_id')->unsigned()->nullable();
             $table->string('ip')->nullable();
-            $table->integer('job_id')->unsigned();
+            $table->bigInteger('job_id')->unsigned();
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
